RMS Import (analysis)

+++ RR-Import/src/main/java/com/scor/rr/repository/rap/TargetRapRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
  * @author Hamiani Mohammed
  * creation date  17/09/2019 at 11:37
  **/
-public interface TargetRapRepository extends JpaRepository<TargetRAP, Long> {
+public interface TargetRapRepository extends JpaRepository<TargetRAP, Integer> {
     List<TargetRAP> findByModellingVendorAndModellingSystemAndModellingSystemVersionAndSourceRapCode(String modellingVendor, String modellingSystem, String modellingSystemVersion, String sourceRapCode);
     TargetRAP findByTargetRAPId(Integer targetRapId);
 }
